Refactor into Swift Package with typography and flowfield styles
Restructure from monolithic script files into a proper SPM package:
- WallpaperKit library: shared rendering, colors, noise, styles
- generate executable: single wallpaper CLI
- setup executable: config-driven batch generation

Each style lives in its own file under Sources/WallpaperKit/Styles/.
setup now calls generateWallpaper() directly instead of shelling out
to a separate Swift process.

New styles:
- typography: scattered text at varying sizes/opacities/rotations,
  three depth layers (giants, mid, detail), design-poster aesthetic
- flowfield: organic flowing lines driven by 2D value noise,
  topographic/wind-map texture, unique pattern per workspace

Task scripts updated to use `swift run generate/setup`.
