+++ .mise/tasks/cli
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 #MISE description="Run generator directly with arguments"
-swift src/generate.swift "$@"
+swift run generate -- "$@"
