+++ .mise/tasks/quick
@@ -17,7 +17,7 @@ SCREEN_W=$(echo "$SCREEN_RES" | cut -d'x' -f1)
 SCREEN_H=$(echo "$SCREEN_RES" | cut -d'x' -f2)
 
 gum spin --spinner dot --title "Generating at ${SCREEN_W}x${SCREEN_H}..." -- \
-    swift src/generate.swift "$NAME" --width "$SCREEN_W" --height "$SCREEN_H"
+    swift run generate -- "$NAME" --width "$SCREEN_W" --height "$SCREEN_H"
 
 gum style --foreground 46 "✅ Generated! Saved to $OUTPUT_DIR"
