+++ Sources/WallpaperKit/Resolutions.swift
@@ -1,0 +1,9 @@
+public let resolutions: [String: (width: Int, height: Int)] = [
+    "1080p": (1920, 1080),
+    "1440p": (2560, 1440),
+    "4k": (3840, 2160),
+    "macbook-14": (3024, 1964),
+    "macbook-16": (3456, 2234),
+    "imac-24": (4480, 2520),
+    "studio-display": (5120, 2880),
+]
