+++ Sources/WallpaperKit/Style.swift
@@ -1,0 +1,8 @@
+public enum WallpaperStyle: String {
+    case classic
+    case diagonal
+    case tiled
+    case flowfield
+    case typography
+    case perspective
+}
